% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaMatMult}
\alias{armaMatMult}
\title{Multiply two matrices using RcppArmadillo}
\usage{
armaMatMult(A, B)
}
\arguments{
\item{A}{B two matrices}
}
\description{
Multiply two matrices using RcppArmadillo
}
