% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Multiply two matrices using RcppEigen::Map}
\usage{
eigenMapMatMult(A, B)
}
\arguments{
\item{A}{B two matrices}
}
\description{
Multiply two matrices using RcppEigen::Map
}
