% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMatMult}
\alias{eigenMatMult}
\title{Multiply two matrices using RcppEigen}
\usage{
eigenMatMult(A, B)
}
\arguments{
\item{A}{B two matrices}
}
\description{
Multiply two matrices using RcppEigen
}
