% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeeds.R
\name{getSeeds}
\alias{getSeeds}
\title{Finds a set of seeds}
\usage{
getSeeds(W, x, s)
}
\arguments{
\item{W}{vector of indices of the shared vertices of \eqn{G_1} \eqn{G_2}}

\item{x}{vector of indices for vertices of interest (voi) in \eqn{G_1}}

\item{s}{number of seeds}
}
\value{
a vector of \eqn{s} seeds for \eqn{sgm}
}
\description{
Finds a set of seeds
}
\author{
Youngser Park <youngser@jhu.edu>
}
