% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass2ranksuplus.r
\name{pass2ranksu}
\alias{pass2ranksu}
\title{Generates a matrix whose \eqn{i,j-th} entry represents the location of
element \eqn{i,j} when the entries in row \eqn{i} are ordered from
 greatest to least. ties broken by averaging.}
\usage{
pass2ranksu(M)
}
\arguments{
\item{M}{A matrix that we wish to pass to ranks on.}
}
\value{
ranked matrix and unique entries sorted
}
\description{
Generates a matrix whose \eqn{i,j-th} entry represents the location of
element \eqn{i,j} when the entries in row \eqn{i} are ordered from
 greatest to least. ties broken by averaging.
}
\author{
Heather Gaddy Patsolic <hgaddy1@jhu.edu>
}
