% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topk.r
\name{topk}
\alias{topk}
\title{Creates top \code{k} nomination list of vertices in \eqn{G_2} for each vertex
     in \eqn{G_1}}
\usage{
topk(k, M)
}
\arguments{
\item{k}{number of vertices we will consider from \eqn{G_2}}

\item{M}{multistart matrix (doubly stochastic)}
}
\value{
probability matrix based on likelihood each vertex in
                \eqn{G2} is matched to corresponding vertex of
                \eqn{G_1} i.e. the \eqn{P[i,:]} gives the
                probabilities of the top \code{k}
                most likely candidates from \eqn{G_2} for vertex
                \eqn{i} in \eqn{G_1}
}
\description{
Creates top \code{k} nomination list of vertices in \eqn{G_2} for each vertex
     in \eqn{G_1}
}
\author{
Heather Gaddy Patsolic <hgaddy1@jhu.edu>
}
